requirements.txt
tabulate
click
cloup
GitPython
pyyaml
google-vizier[jax]
pandas
matplotlib
